Move `load_image_from_url` helper to `utils` and clean up imports
- Extract the `load_image_from_url` function from `src/webui/plantimager/webui/pages/table.py` and place it in `src/webui/plantimager/webui/utils.py`
- Remove the duplicate function and its imports from `table.py`
- Add the missing `base64` and `requests` imports to `utils.py`
- Update imports in `visu.py` to use the helper from `utils`
- Remove unused imports (`base64`, `requests`) from `table.py` and adjust remaining imports
- Ensure calls to `load_image_from_url` in the table and carousel now reference the new location in `utils`
