+++ src/webui/plantimager/webui/utils.py
@@ -23,9 +23,10 @@ Usage Examples
 >>> has_config = has_pipeline_cfg('localhost', '5000', 'my_plant_scan')
 ```
 """
-
+import base64
 from typing import Any
 
+import requests
 from dash import dcc
 from plantdb.client.rest_api import parse_scans_info
 from plantdb.client.rest_api import request_check_username
@@ -143,3 +144,35 @@ def _validate_new_username(username: str, host: str, port: str, prefix: str, ssl
             return True  # Available username
     except Exception as e:
         return False
+
+
+def load_image_from_url(url):
+    """Load an image from a given URL and encode it to a base64 data URI.
+
+    Parameters
+    ----------
+    url : str
+        The base URL to which the function will append itself in order to form
+        the full request URL.
+
+    Returns
+    -------
+    str
+        A base64‑encoded data URI containing the image, or a fallback string
+        such as "No Image" or "Error Loading" when the fetch is unsuccessful.
+    """
+    # Fetch image and convert to base64
+    try:
+        full_url = f"{url}{url}"
+        response = requests.get(full_url)
+        if response.status_code == 200:
+            content_type = response.headers.get('content-type', 'image/jpeg')
+            if not content_type.startswith('image'):
+                return "Not an image"
+            encoded_img = base64.b64encode(response.content).decode('ascii')
+            img_data = f"data:{content_type};base64,{encoded_img}"
+        else:
+            img_data = "No Image"
+    except Exception:
+        img_data = "Error Loading"
+    return img_data
